Move tests, fix type hints and imports
